docker/eto-webapp/requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
psycopg2-binary==2.9.9
redis==5.0.1
SQLAlchemy==2.0.23

# Environment & Configuration
python-dotenv==1.0.0
PyYAML==6.0.1
configparser==6.0.0

# Monitoring & Metrics
prometheus-flask-exporter==0.22.4
py-healthcheck==1.10.1
statsd==4.0.1
datadog==0.47.0

# Security
Flask-Security-Too==5.3.2
PyJWT==2.8.0
cryptography==41.0.5
passlib==1.7.4

# Performance & Caching
Flask-Caching==2.1.0
uWSGI==2.0.23
gunicorn==21.2.0

# Logging & Tracing
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2

# Development Tools
black==23.10.1
flake8==6.1.0
mypy==1.6.1
isort==5.12.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
